decouple Song from MidiCompiler, move api/ layer
Song is now a pure data container with zero compiler
references. All deprecated methods removed (generate,
save, midigen property, available_channels).

MidiCompiler imports Song at top level — no more
late-import circular dependency hack.
