+++ midigen/api/__init__.py
@@ -1,1 +1,5 @@
 """High-level API layer — depends on all layers."""
+
+from midigen.api.song import Song
+from midigen.api.midigen import MidiGen
+from midigen.api.compiler import MidiCompiler
